+++ test/integration/src/segments/ConcatRandomTimeEchoSegment.js
@@ -19,31 +19,51 @@ export default class ConcatRandomTimeEchoSegment extends MapSegment {
   }
 
   _generateQueryId(query) {
-    return query.value + (query.isParallel ? '$p' : '$s');
+    return query.value + (query.isParallel ? '$p' : '$s') +
+      (query.isReplaceParallel ? '$p' : '$s') +
+      (query.shouldReplace ? '$r' : '');
   }
 
   _generateThunkFunction(thunk) {
     var query = thunk.query;
     var queryId = thunk.queryId;
-    var dependencies, i;
+    var dependencies, recursiveDependencies, RecursiveQueryCtor, i, val;
 
-    if (query.isParallel) {
-      dependencies = QueryDependencies.parallel(Promise);
-    } else {
-      dependencies = QueryDependencies.serial(Promise);
-    }
+    dependencies = query.isParallel ?
+      QueryDependencies.parallel(Promise) : QueryDependencies.serial(Promise);
+    RecursiveQueryCtor = query.isReplaceParallel ?
+      QueryDependencies.parallel : QueryDependencies.serial;
 
     for (i = 0; i < query.value.length; i++) {
+      val = query.value[i];
+      if (query.shouldReplace && val == 's') {
+        recursiveDependencies = new RecursiveQueryCtor(Promise);
+        recursiveDependencies.add('s', RandomTimeEchoSegment.id(), {value: 's'}, true);
+        recursiveDependencies.add('o', RandomTimeEchoSegment.id(), {value: 'o'}, true);
+        recursiveDependencies.add('y', RandomTimeEchoSegment.id(), {value: 'y'}, true);
+        recursiveDependencies.add('a', RandomTimeEchoSegment.id(), {value: 'a'}, true);
+        dependencies.addRecursive(i + '', recursiveDependencies);
+        continue;
+      }
+
       // Always do force load so that response time is always random.
-      dependencies.add(i + '', RandomTimeEchoSegment.id(), {value: query.value[i]}, true);
+      dependencies.add(i + '', RandomTimeEchoSegment.id(), {value: val}, true);
     }
 
     thunk.dependencies = dependencies;
     thunk.func = (dispatch) => {
       // Put results to array, we're going to sort them by the time they arrive.
-      var i, resultArray = [];
+      var i, result, resultArray = [];
       for (i = 0; i < query.value.length; i++) {
-        resultArray.push(dependencies.getResult(i + ''));
+        result = dependencies.getResult(i + '');
+        if (result instanceof QueryDependencies) {
+          resultArray.push(result.getResult('s'));
+          resultArray.push(result.getResult('o'));
+          resultArray.push(result.getResult('y'));
+          resultArray.push(result.getResult('a'));
+          continue;
+        }
+        resultArray.push(result);
       }
       resultArray.sort(function(a, b) {
         if (a.updated == b.updated) return 0;